cleanup

+++ scripts/python/finance_api/stockprices.py
@@ -1,6 +1,6 @@
 # %%
 
-# Script to get stock information on companies via yfinance API 
+# Script to get stock information on companies via yfinance API
 # https://pypi.org/project/yfinance/
 
 import plotly.graph_objects as go
@@ -8,33 +8,34 @@ from datetime import datetime
 import pandas as pd
 import yfinance as yf
 
-print('Enter your ticker:')
+print("Enter your ticker:")
 x = input()
 
-print('Enter your period? (1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y, 10y, ytd, max:)')
+print("Enter your period? (1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y, 10y, ytd, max:)")
 y = input()
 
 df = yf.Ticker(x)
 
 # show analysts recommendations
 recom = df.recommendations
-final_df = recom.sort_values(by=['Date'], ascending=False).head(5)
+final_df = recom.sort_values(by=["Date"], ascending=False).head(5)
 print(final_df)
 
 
-old = df.history(period = y)
+old = df.history(period=y)
 old.head()
 
 
 # Reset index of dataframe in order to visulaise
 old = old.reset_index()
-for i in ['Open', 'High', 'Close', 'Low']: 
-      old[i]  =  old[i].astype('float64')
+for i in ["Open", "High", "Close", "Low"]:
+    old[i] = old[i].astype("float64")
 
 
-# Plot linechart for period 
+# Plot linechart for period
 import plotly.express as px
-fig = px.line(old, x="Date", y="Open", title= x + ' Stock Prices')
+
+fig = px.line(old, x="Date", y="Open", title=x + " Stock Prices")
 fig.show()
